requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-telegram-bot>=21.0
openai>=1.6.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pydantic>=2.5.0
httpx>=0.25.0
gdown>=5.1.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
duckduckgo-search>=7.0.0
